# Review round 2 — address P2-1, P2-2, P2-3, P3 + recommended tests
## P2-1 — Architecture test was vacuous (fixed + mutation-verified)

`TramaiInvocationHandlerArchitectureTest` compared fully-qualified `Class.name`
strings against bare simple names, so a forbidden coordinator field would never
fail the test. Now compares actual `Class<?>` objects:

```kotlin
val fieldTypes = TramaiInvocationHandler::class.java.declaredFields.map { it.type }.toSet()
...
assertThat(fieldTypes).doesNotContain(ProviderExecutionCoordinator::class.java)
```

Mutation-verified: temporarily adding a `RawResponseCoordinator` field to the
handler makes `handler owns no execution algorithms` FAIL; reverting restores
green.

## P2-2 — InvocationExecutionCoordinator was becoming the new god-object (fixed)

The coordinator's KDoc claimed it "does not implement provider, tool, approval,
DLP, cache, memory, or streaming algorithms", but it directly owned the DLP
scan (`inspectDlpAuthoritatively`, `sanitizeProviderResponse`,
`applyProviderOutputDlp`) and the entire post-claim approval subsystem
(`executeResumedTool`, `continueAfterToolResult`,
`processRemainingResumeToolCalls`, `appendRemainingResumeToolResult`,
`finalizeResumedOperation`).

Two new narrow collaborators, both pure verbatim moves:

- **`ProviderResponseDlpSanitizer`** (new file, 138 lines): authoritative
  model-output DLP scan, consistency validation, redaction-audit emission,
  failure mapping to `DlpInspectionException`. Wired into the parent's
  `ProviderResponseSanitizer` lambda.
- **`ClaimedResumeExecutionCoordinator`** (new file, 351 lines): the
  post-claim approval path — resumed tool execution, sibling-tool processing,
  reinjection, memory persistence, policy checks, observation completion,
  return-kind finalization. The parent's `execute(ClaimedResumeExecutionRequest)`
  is now a one-line delegate.

`InvocationExecutionCoordinator` is now ~400 lines (was 789): builds the
execution graph, holds the policy gates, dispatches by return kind, and
delegates DLP + claimed-resume. This matches the roadmap rule "the top-level
invocation coordinator contains only execution sequencing."

## P2-3 — Roadmap was stale (fixed)

`ROADMAP-0.6.0.md`: Epic 3.7 marked ✅ COMPLETE — PR #245; Phase 3 explicitly
marked complete before Phase 4.

## P3 — Dead DLP helper (fixed)

`inspectDlpForDetectionOnly` removed with the DLP extraction (it was only in
the coordinator; `ToolResultSanitizer`'s copy is separate and used).

## Recommended tests (ToolLoopCoordinator)

New `ToolLoopCoordinatorTest` (6 tests, mutation-sensitive):
- cancellation during tool reinjection → `onCallCancelled`, not `onCallCompleted`
- unknown tool call → normalized to `unregistered_tool` placeholder
- token budget failure → completes observation + throws
- max 5 iterations → `Exceeded maximum tool call loops (5)`
- ordinary reinjection error (policy denial) → primary error preserved + observation completes
- no tool calls → returns provider result directly

Skipped a dedicated `InvocationExecutionCoordinatorTest`: the three suggested
assertions (one delegate per ReturnKind, UNIT returns Unit, plan fingerprint
forwarded) are already exercised by the 589-test engine suite + 20
byte-identical characterization traces that run through the real coordinator;
a dedicated test would duplicate the engine harness without new signal.

## Verification

```
./gradlew :tramai-engine:test :tramai-engine:apiCheck \
  verifyCancellationSafety verifyChangePolicy verifyMaintainabilityBaseline \
  -PchangeClass=runtime-behaviour --no-daemon
```

- BUILD SUCCESSFUL in 2m35s
- Maintainability baseline PASSED (12 accepted deviations; MQ-0017 unchanged)
- 20/20 characterization traces byte-identical (0 modified in git)
- apiCheck: zero public API diff
- P1/P2/P3 review findings addressed; no behavior change intended
